+++ mitra_activitypub/src/handlers/undo.rs
@@ -110,7 +110,7 @@ pub async fn handle_undo(
 
     match get_reaction_by_remote_activity_id(db_client, &activity.object).await {
         Ok(reaction) => {
-            // Undo(Like) or Undo(EmojiReact)
+            // Undo(Like), Undo(EmojiReact), Undo(Dislike)
             if reaction.author_id != actor_profile.id {
                 return Err(ValidationError("actor is not an author").into());
             };