+++ CHANGELOG.md
@@ -12,6 +12,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 - Use actor ID as handle if acct URI is not known.
 - Added `enable-fep-ef61` command.
 - Support unicode emoji shortcodes.
+- Process incoming `Dislike` activities as "👎" reaction.
 
 ### Changed